# packages/api-client/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "pmagent API",
...
    "version": "0.1.0"
  },
  "paths": {
...
    "/v1/workspaces/{workspace_id}/invites": {
      "get": {
        "tags": [
          "invites"
        ],
        "summary": "List Invites",
        "description": "Pending invites: not accepted, revoked, expired, or used up. Owners and admins.",
        "operationId": "list_invites",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "workspace_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Workspace Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/InviteRead"
                  },
                  "title": "Response List Invites"
                }
              }
            }
          },
          "401": {
            "description": "Missing, invalid, or expired credentials",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "403": {
            "description": "Signed in, but your role or token scope doesn't allow this",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "404": {
            "description": "Not found, or not visible to you",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "invites"
        ],
        "summary": "Invite By Email",
        "description": "Email an invite (valid 7 days, single use, only for that address). Inviting the same\naddress again replaces the earlier invite. Owners and admins.",
        "operationId": "invite_by_email",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "workspace_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Workspace Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EmailInviteCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InviteRead"
                }
              }
            }
          },
          "401": {
            "description": "Missing, invalid, or expired credentials",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "403": {
            "description": "Signed in, but your role or token scope doesn't allow this",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "404": {
            "description": "Not found, or not visible to you",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "409": {
            "description": "Conflicts with the current state",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "422": {
            "description": "Request body or parameters failed validation",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          }
        }
      }
    },
    "/v1/workspaces/{workspace_id}/invites/links": {
      "post": {
        "tags": [
          "invites"
        ],
        "summary": "Create Invite Link",
        "description": "Create a shareable invite link (member or guest only). The `url` is returned only\nhere. Owners and admins.",
        "operationId": "create_invite_link",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "workspace_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Workspace Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LinkInviteCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LinkInviteCreated"
                }
              }
            }
          },
          "401": {
            "description": "Missing, invalid, or expired credentials",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "403": {
            "description": "Signed in, but your role or token scope doesn't allow this",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "404": {
            "description": "Not found, or not visible to you",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "409": {
            "description": "Conflicts with the current state",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "422": {
            "description": "Request body or parameters failed validation",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          }
        }
      }
    },
    "/v1/workspaces/{workspace_id}/invites/{invite_id}": {
      "delete": {
        "tags": [
          "invites"
        ],
        "summary": "Revoke Invite",
        "description": "Revoke an email invite or link. Owners and admins.",
        "operationId": "revoke_invite",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "invite_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Invite Id"
            }
          },
          {
            "name": "workspace_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Workspace Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "401": {
            "description": "Missing, invalid, or expired credentials",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "403": {
            "description": "Signed in, but your role or token scope doesn't allow this",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "404": {
            "description": "Not found, or not visible to you",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          }
        }
      }
    },
    "/v1/invites/preview": {
      "post": {
        "tags": [
          "invites"
        ],
        "summary": "Preview Invite",
        "description": "No sign-in needed: the accept page shows this before sign-in or sign-up.",
        "operationId": "preview_invite",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AcceptInvite"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InvitePreview"
                }
              }
            }
          },
          "400": {
            "description": "Bad request, for example an invalid or expired link or code",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "422": {
            "description": "Request body or parameters failed validation",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          }
        }
      }
    },
    "/v1/invites/accept": {
      "post": {
        "tags": [
          "invites"
        ],
        "summary": "Accept Invite",
        "description": "Join the workspace. Email invites must be accepted by the invited address (403\notherwise). If you're already a member, your role doesn't change.",
        "operationId": "accept_invite",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AcceptInvite"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkspaceWithRole"
                }
              }
            }
          },
          "400": {
            "description": "Bad request, for example an invalid or expired link or code",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "401": {
            "description": "Missing, invalid, or expired credentials",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "403": {
            "description": "Signed in, but your role or token scope doesn't allow this",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "422": {
            "description": "Request body or parameters failed validation",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    }
  },
  "components": {
    "schemas": {
      "AcceptInvite": {
        "properties": {
          "token": {
            "type": "string",
            "maxLength": 256,
            "title": "Token"
          }
        },
        "type": "object",
        "required": [
          "token"
        ],
        "title": "AcceptInvite"
      },
...
      "EmailInviteCreate": {
        "properties": {
          "email": {
            "type": "string",
            "format": "email",
            "title": "Email"
          },
          "role": {
            "type": "string",
            "enum": [
              "admin",
              "member",
              "guest"
            ],
            "title": "Role",
            "default": "member"
          }
        },
        "type": "object",
        "required": [
          "email"
        ],
        "title": "EmailInviteCreate"
      },
      "InviteKind": {
        "type": "string",
        "enum": [
          "email",
          "link"
        ],
        "title": "InviteKind"
      },
      "InvitePreview": {
        "properties": {
          "workspace_name": {
            "type": "string",
            "title": "Workspace Name"
          },
          "workspace_kind": {
            "$ref": "#/components/schemas/WorkspaceKind"
          },
          "role": {
            "$ref": "#/components/schemas/Role"
          },
          "invited_by": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Invited By"
          },
          "email": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Email"
          },
          "expires_at": {
            "type": "string",
            "format": "date-time",
            "title": "Expires At"
          }
        },
        "type": "object",
        "required": [
          "workspace_name",
          "workspace_kind",
          "role",
          "invited_by",
          "email",
          "expires_at"
        ],
        "title": "InvitePreview",
        "description": "What the accept page shows before the user signs in or accepts."
      },
      "InviteRead": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "kind": {
            "$ref": "#/components/schemas/InviteKind"
          },
          "email": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Email"
          },
          "role": {
            "$ref": "#/components/schemas/Role"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "expires_at": {
            "type": "string",
            "format": "date-time",
            "title": "Expires At"
          },
          "max_uses": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Max Uses"
          },
          "use_count": {
            "type": "integer",
            "title": "Use Count"
          }
        },
        "type": "object",
        "required": [
          "id",
          "kind",
          "email",
          "role",
          "created_at",
          "expires_at",
          "max_uses",
          "use_count"
        ],
        "title": "InviteRead"
      },
      "LinkInviteCreate": {
        "properties": {
          "role": {
            "type": "string",
            "enum": [
              "member",
              "guest"
            ],
            "title": "Role",
            "default": "member"
          },
          "max_uses": {
            "anyOf": [
              {
                "type": "integer",
                "maximum": 1000.0,
                "minimum": 1.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Max Uses"
          },
          "expires_in_days": {
            "type": "integer",
            "maximum": 30.0,
            "minimum": 1.0,
            "title": "Expires In Days",
            "default": 7
          }
        },
        "type": "object",
        "title": "LinkInviteCreate"
      },
      "LinkInviteCreated": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "kind": {
            "$ref": "#/components/schemas/InviteKind"
          },
          "email": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Email"
          },
          "role": {
            "$ref": "#/components/schemas/Role"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "expires_at": {
            "type": "string",
            "format": "date-time",
            "title": "Expires At"
          },
          "max_uses": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Max Uses"
          },
          "use_count": {
            "type": "integer",
            "title": "Use Count"
          },
          "url": {
            "type": "string",
            "title": "Url"
          }
        },
        "type": "object",
        "required": [
          "id",
          "kind",
          "email",
          "role",
          "created_at",
          "expires_at",
          "max_uses",
          "use_count",
          "url"
        ],
        "title": "LinkInviteCreated",
        "description": "Returned once, at creation: the only time the link's token is visible."
      },
...
      "ProblemDetail": {
        "properties": {
          "type": {
            "type": "string",
            "title": "Type",
            "examples": [
              "https://pmagent.dev/problems/not_found"
            ]
          },
          "title": {
            "type": "string",
            "title": "Title",
            "examples": [
              "Not Found"
            ]
          },
          "status": {
            "type": "integer",
            "title": "Status",
            "examples": [
              404
            ]
          },
          "detail": {
            "type": "string",
            "title": "Detail",
            "examples": [
              "Workspace not found"
            ]
          },
          "request_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Request Id",
            "examples": [
              "3f2b9c0e8a4d4e8f9b1c2d3e4f5a6b7c"
            ]
          },
          "errors": {
            "anyOf": [
              {
                "items": {
                  "additionalProperties": true,
                  "type": "object"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Errors",
            "description": "Field errors; only on 422 validation errors."
          }
        },
        "type": "object",
        "required": [
          "type",
          "title",
          "status",
          "detail"
        ],
        "title": "ProblemDetail",
        "description": "RFC 9457 problem details. Every error response has this shape."
      },
...
      "Role": {
        "type": "string",
        "enum": [
          "owner",
          "admin",
          "member",
          "guest"
        ],
        "title": "Role"
      },
...
      "WorkspaceKind": {
        "type": "string",
        "enum": [
          "personal",
          "team",
          "business"
        ],
        "title": "WorkspaceKind"
      },
...
      "WorkspaceWithRole": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "slug": {
            "type": "string",
            "title": "Slug"
          },
          "kind": {
            "$ref": "#/components/schemas/WorkspaceKind"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "role": {
            "$ref": "#/components/schemas/Role"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "slug",
          "kind",
          "created_at",
          "role"
        ],
        "title": "WorkspaceWithRole"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "description": "Access token from /v1/auth/login, or an API token (pmat_...)",
        "scheme": "bearer"
      }
    }
  },
...
}